requirements.txt
telebot
openai
langchain
python-dotenv
pypdf
chromadb
tiktoken
tabulate